requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2.3
djangorestframework==3.12.4
idna==2.10
Pillow==8.2.0
pytz==2021.1
qrcode==6.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
